chore: move Balance out of keychain module
