+++ crates/chain/tests/test_tx_graph_conflicts.rs
@@ -3,7 +3,7 @@ mod common;
 
 use std::collections::{BTreeSet, HashSet};
 
-use bdk_chain::{keychain::Balance, BlockId};
+use bdk_chain::{Balance, BlockId};
 use bitcoin::{OutPoint, Script};
 use common::*;
